requirements.txt
paho-mqtt==1.2.3
psutil==5.2.2
pyserial==3.3
